feat: AWS Lambda対応のためserverless-expressを導入
Expressアプリ定義をapp.jsへ分離し、index.jsはローカル開発用の
薄いエントリーポイント（AWS_LAMBDA_FUNCTION_NAME未設定時のみ
app.listen()）に変更。lambda.jsでserverless-expressを使い
appをLambdaハンドラーとしてラップする。

あわせて、app.js作成時に#59のCORS Authorizationヘッダー修正
(develop未マージ)が失われていたため、app.js側にも同じ修正を適用。

Fixes #41
