+++ backend/package.json
@@ -26,6 +26,7 @@
   },
   "dependencies": {
     "@prisma/client": "^6.6.0",
+    "@vendia/serverless-express": "^4.12.6",
     "bcryptjs": "^3.0.2",
     "dotenv": "^16.4.7",
     "express": "^4.21.2",
